requirements.txt
requests
dnspython
python-whois
